refactor: home 패키지명을 maps로 변경
