+++ android/Staccato_AN/app/src/main/java/com/woowacourse/staccato/presentation/maps/MapsViewModelFactory.kt
@@ -1,4 +1,4 @@
-package com.woowacourse.staccato.presentation.home
+package com.woowacourse.staccato.presentation.maps
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
