+++ android/Staccato_AN/app/src/main/res/layout/activity_main.xml
@@ -22,7 +22,7 @@
 
             <androidx.fragment.app.FragmentContainerView
                 android:id="@+id/fragment_container_view_map"
-                android:name="com.woowacourse.staccato.presentation.home.MapsFragment"
+                android:name="com.woowacourse.staccato.presentation.maps.MapsFragment"
                 android:layout_width="0dp"
                 android:layout_height="0dp"
                 app:layout_constraintBottom_toBottomOf="parent"
